requirements.txt
Django==1.8.2
Markdown==2.6.2
django-extensions==1.5.5
django-filter==0.10.0
djangorestframework==3.1.2
ipython==3.1.0
six==1.9.0
django-cors-headers==1.1.0
